kubi_callcenter_zusagen/generate_template/requirements.txt
paramiko
pandas
xlrd
openpyxl
didas[oracle]
python-dotenv